refactor: hook 의 onInputChange 로직을 container 컴포넌트로 이동
