+++ src/containers/CardOwnerInput/CardOwnerInput.jsx
@@ -1,9 +1,21 @@
 import useCardOwner from "../../hooks/useCardOwner";
 import { CARD_INPUT, INPUT_LABEL_TEXT } from "../../constants";
 import TextLimitInput from "../../components/Input/TextLimitInput/TextLimitInput";
+import { isNumberText } from "../../utils/cardInputValidation";
 
 const CardOwnerInput = ({ className }) => {
-  const { cardOwnerState, onCardOwnerInputChange } = useCardOwner();
+  const { cardOwnerState, setCardOwnerState } = useCardOwner();
+
+  const onCardOwnerInputChange = (event) => {
+    const { value } = event.target;
+
+    if (isNumberText(value) && value !== "") {
+      event.target.value = event.target.value.slice(0, -1);
+      return;
+    }
+
+    setCardOwnerState(value);
+  };
 
   return (
     <TextLimitInput
