+++ src/containers/CardPasswordInput/CardPasswordInput.jsx
@@ -1,9 +1,22 @@
 import { CARD_INPUT, INPUT_LABEL_TEXT } from "../../constants";
 import InputBoxList from "../../components/Input/InputBoxList/InputBoxList";
 import useCardPassword from "../../hooks/useCardPassword";
+import { isNumberText } from "../../utils/cardInputValidation";
 
 const CardPasswordInput = ({ className }) => {
-  const { onCardPasswordInputChange } = useCardPassword();
+  const { cardPassword, setCardPasswordState } = useCardPassword();
+
+  const onCardPasswordInputChange = (event) => {
+    const { value, name: inputIndex } = event.target;
+    if (!isNumberText(value)) {
+      event.target.value = event.target.value.slice(0, -1);
+      return;
+    }
+
+    const newCardPassword = [...cardPassword];
+    newCardPassword[Number(inputIndex)] = value;
+    setCardPasswordState(newCardPassword);
+  };
 
   return (
     <InputBoxList
