+++ src/hooks/useCardCompany.js
@@ -1,7 +1,6 @@
 import { useCallback, useContext } from "react";
 import { STATE_KEY } from "../constants";
 import AppContext from "../contexts/appContext";
-import { isNumberText } from "../utils/cardInputValidation";
 
 const useCardCompany = () => {
   const { setCardStateByKey } = useContext(AppContext);
@@ -13,23 +12,8 @@ const useCardCompany = () => {
     [setCardStateByKey]
   );
 
-  const onCardCompanyInputChange = useCallback(
-    (event) => {
-      const { value } = event.target;
-
-      if (!isNumberText(value)) {
-        event.target.value = event.target.value.slice(0, -1);
-        return;
-      }
-
-      setCardCompanyState(value);
-    },
-    [setCardCompanyState]
-  );
-
   return {
     setCardCompanyState,
-    onCardCompanyInputChange,
   };
 };
