+++ src/hooks/useCardExpiration.js
@@ -1,7 +1,6 @@
 import { useCallback, useContext } from "react";
 import { STATE_KEY } from "../constants";
 import AppContext from "../contexts/appContext";
-import { isNumberText } from "../utils/cardInputValidation";
 
 const useCardExpiration = () => {
   const { cardState, setCardStateByKey } = useContext(AppContext);
@@ -15,26 +14,9 @@ const useCardExpiration = () => {
     [setCardStateByKey]
   );
 
-  const onCardExpirationInputChange = useCallback(
-    (event) => {
-      const { value, name } = event.target;
-
-      if (!isNumberText(value)) {
-        event.target.value = event.target.value.slice(0, -1);
-        return;
-      }
-
-      setCardExpirationState({
-        ...cardExpirationState,
-        [name]: value,
-      });
-    },
-    [cardExpirationState, setCardExpirationState]
-  );
-
   return {
     cardExpirationState,
-    onCardExpirationInputChange,
+    setCardExpirationState,
   };
 };
