+++ src/hooks/useCardOwner.js
@@ -1,7 +1,6 @@
 import { useCallback, useContext } from "react";
 import { STATE_KEY } from "../constants";
 import AppContext from "../contexts/appContext";
-import { isNumberText } from "../utils/cardInputValidation";
 
 const useCardOwner = () => {
   const { cardState, setCardStateByKey } = useContext(AppContext);
@@ -15,23 +14,9 @@ const useCardOwner = () => {
     [setCardStateByKey]
   );
 
-  const onCardOwnerInputChange = useCallback(
-    (event) => {
-      const { value } = event.target;
-
-      if (isNumberText(value) && value !== "") {
-        event.target.value = event.target.value.slice(0, -1);
-        return;
-      }
-
-      setCardOwnerState(value);
-    },
-    [setCardOwnerState]
-  );
-
   return {
     cardOwnerState,
-    onCardOwnerInputChange,
+    setCardOwnerState,
   };
 };
