+++ src/sections/CardInputSection/CardInputSection.jsx
@@ -22,6 +22,7 @@ const CardInputSection = ({ showCardCompanySelectSection }) => {
   const { setCardCompanyState } = useCardCompany();
 
   useEffect(() => {
+    console.log(cardNumberState);
     if (!isAllTextFilledInObject(cardNumberState, CARD_INPUT.CARD_NUMBER_TEXT_LENGTH)) {
       return;
     }
